requirements.txt
# Dependencias core
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=4.9.0

# Supabase - usar version sin pyiceberg
supabase==2.3.0

# FastAPI y servidor
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# CORS
python-multipart>=0.0.6
